MAR-1961 Added null check

+++ projects/FieldNation/fieldnation/src/main/java/com/fieldnation/service/auth/AuthSystem.java
@@ -373,7 +373,7 @@ public class AuthSystem implements AuthTopicConstants {
                     _account = null;
                     _accountManager.invalidateAuthToken(getAccountType(), bundle.getString(AccountManager.KEY_AUTHTOKEN));
                     setState(AuthState.NOT_AUTHENTICATED);
-                    _authToken.delete();
+                    if (_authToken != null) _authToken.delete();
                     _authToken = null;
                     return;
                 }
